winning

+++ db/migrate/20151213035429_create_searches.rb
@@ -14,8 +14,8 @@ class CreateSearches < ActiveRecord::Migration
       t.float :gpa
       t.integer :workexperience 
       t.string :livingin 
-      t.integer :monthsspentabroadliving
-      t.integer :monthsspentabroadworking
+      t.float :monthsspentabroadliving
+      t.float :monthsspentabroadworking
       t.string :degree 
       t.string :degreetype 
       t.string :countryofdegree
